feat: changes to support SQLAlchemy >= 2, and Pandas >= 2
To support SQLAlchemy 2 (which seems to also then require Pandas >= 2 for its
read_sql function) we need to:

For SQLAlchemy >= 2
- Always wrap SQL strings with the `text` method
- Use named parameters in SQL strings with dictionaries as the values of the
  parameters.

For Pandas >= 2
- Connect to the database outside of the pandas read_sql function in our
  own context
- Use pd.testing.assert_frame_equal rather than pd.util.testing.assert_frame_equal

Doing this in the one change rather than two because of the interplay between
SQLAlchemy and Pandas in read_sql.

This is part of work of getting https://github.com/uktrade/data-store-service
to build with the recent changes of
https://github.com/uktrade/data-engineering-common/pull/34 that bumped
SQLAlchemy to >= 2. The changes here fix a number of errors in
data-store-service's testing suite.

To actually test the data-store-service before merge of this change, its
requirements.txt file had to be changed to refer to the specific commit in this
PR, for example:

data-engineering-common @ git+https://github.com/uktrade/data-engineering-common.git@COMMIT_ID

(replacing `COMMIT_ID` with the actual commit being tested)